MatchInfo subclasses RelationInfo, CaptureInfo.
