+++ engine/src/main/java/nl/inl/blacklab/search/BlackLabIndexIntegrated.java
@@ -40,7 +40,7 @@ import nl.inl.blacklab.search.indexmetadata.IndexMetadataIntegrated;
 import nl.inl.blacklab.search.indexmetadata.IndexMetadataWriter;
 import nl.inl.blacklab.search.indexmetadata.RelationUtil;
 import nl.inl.blacklab.search.lucene.BLSpanQuery;
-import nl.inl.blacklab.search.lucene.MatchInfo;
+import nl.inl.blacklab.search.lucene.RelationInfo;
 import nl.inl.blacklab.search.lucene.SpanQueryRelations;
 import nl.inl.blacklab.search.results.QueryInfo;
 
@@ -186,12 +186,12 @@ public class BlackLabIndexIntegrated extends BlackLabIndexAbstract {
     public BLSpanQuery tagQuery(QueryInfo queryInfo, String luceneField, String tagName,
             Map<String, String> attributes) {
         return relationQuery(queryInfo, luceneField, RelationUtil.inlineTagFullType(tagName),
-                attributes, SpanQueryRelations.Direction.BOTH_DIRECTIONS, MatchInfo.SpanMode.FULL_SPAN);
+                attributes, SpanQueryRelations.Direction.BOTH_DIRECTIONS, RelationInfo.SpanMode.FULL_SPAN);
     }
 
     @Override
     public BLSpanQuery relationQuery(QueryInfo queryInfo, String luceneField, String relationType,
-            Map<String, String> attributes, SpanQueryRelations.Direction direction, MatchInfo.SpanMode spanMode) {
+            Map<String, String> attributes, SpanQueryRelations.Direction direction, RelationInfo.SpanMode spanMode) {
         return new SpanQueryRelations(queryInfo, luceneField, relationType, attributes, direction,
                 spanMode);
     }
