+++ engine/src/main/java/nl/inl/blacklab/search/lucene/PerDocumentSortedSpans.java
@@ -179,7 +179,7 @@ final class PerDocumentSortedSpans extends BLFilterDocsSpans<SpansInBuckets> {
     }
 
     @Override
-    public MatchInfo getRelationInfo() {
+    public RelationInfo getRelationInfo() {
         return in.getRelationInfo(indexInBucket);
     }
