+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpanQueryRelationSpanAdjust.java
@@ -21,7 +21,7 @@ import nl.inl.blacklab.search.results.QueryInfo;
  */
 public class SpanQueryRelationSpanAdjust extends BLSpanQuery {
 
-    public static SpanGuarantees createGuarantees(SpanGuarantees clause, MatchInfo.SpanMode mode) {
+    public static SpanGuarantees createGuarantees(SpanGuarantees clause, RelationInfo.SpanMode mode) {
         // NOTE: we don't know the direction here, so choose the one that gives the fewest guarantees
         //   (maybe SpanGuarantees could have a guarantee to help with this? Or maybe not worth it just for relations)
         SpanQueryRelations.Direction direction = SpanQueryRelations.Direction.BOTH_DIRECTIONS;
@@ -30,9 +30,9 @@ public class SpanQueryRelationSpanAdjust extends BLSpanQuery {
 
     private final BLSpanQuery clause;
 
-    private final MatchInfo.SpanMode mode;
+    private final RelationInfo.SpanMode mode;
 
-    public SpanQueryRelationSpanAdjust(BLSpanQuery clause, MatchInfo.SpanMode mode) {
+    public SpanQueryRelationSpanAdjust(BLSpanQuery clause, RelationInfo.SpanMode mode) {
         super(clause.queryInfo);
         this.clause = clause;
         this.mode = mode;
