+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansInBucketsAbstract.java
@@ -44,7 +44,7 @@ abstract class SpansInBucketsAbstract extends SpansInBuckets {
     /**
      * For each hit we fetched, store the active relation info, if any.
      */
-    protected ObjectArrayList<MatchInfo> activeRelationPerHit = null;
+    protected ObjectArrayList<RelationInfo> activeRelationPerHit = null;
 
     private HitQueryContext hitQueryContext;
 
@@ -212,7 +212,7 @@ abstract class SpansInBucketsAbstract extends SpansInBuckets {
     }
 
     @Override
-    public MatchInfo getRelationInfo(int indexInBucket) {
+    public RelationInfo getRelationInfo(int indexInBucket) {
         if (!doMatchInfo)
             return null;
         return activeRelationPerHit.get(indexInBucket);
