+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansInBucketsSameStartEnd.java
@@ -36,7 +36,7 @@ class SpansInBucketsSameStartEnd extends SpansInBuckets {
     /**
      * For each hit we fetched, store the active relation info, if any.
      */
-    protected ObjectArrayList<MatchInfo> activeRelationPerHit = null;
+    protected ObjectArrayList<RelationInfo> activeRelationPerHit = null;
 
     private HitQueryContext hitQueryContext;
 
@@ -189,7 +189,7 @@ class SpansInBucketsSameStartEnd extends SpansInBuckets {
     }
 
     @Override
-    public MatchInfo getRelationInfo(int indexInBucket) {
+    public RelationInfo getRelationInfo(int indexInBucket) {
         return doMatchInfo ? activeRelationPerHit.get(indexInBucket) : null;
     }
