+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansNGrams.java
@@ -233,7 +233,7 @@ class SpansNGrams extends BLSpans {
     }
 
     @Override
-    public MatchInfo getRelationInfo() {
+    public RelationInfo getRelationInfo() {
         return null;
     }
