+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansNot.java
@@ -293,7 +293,7 @@ class SpansNot extends BLSpans {
     }
 
     @Override
-    public MatchInfo getRelationInfo() {
+    public RelationInfo getRelationInfo() {
         return null;
     }
