+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansRelationSpanAdjust.java
@@ -11,7 +11,7 @@ import org.apache.lucene.search.spans.FilterSpans;
 class SpansRelationSpanAdjust extends BLFilterSpans<BLSpans> {
 
     /** how to adjust spans */
-    private final MatchInfo.SpanMode mode;
+    private final RelationInfo.SpanMode mode;
     
     /** Adjusted start position of current hit */
     private int startAdjusted = -1;
@@ -29,14 +29,14 @@ class SpansRelationSpanAdjust extends BLFilterSpans<BLSpans> {
      * @param in spans to adjust
      * @param mode how to adjust spans
      */
-    public SpansRelationSpanAdjust(BLSpans in, MatchInfo.SpanMode mode) {
+    public SpansRelationSpanAdjust(BLSpans in, RelationInfo.SpanMode mode) {
         super(in, SpanQueryRelationSpanAdjust.createGuarantees(in.guarantees(), mode));
         this.mode = mode;
     }
 
     @Override
     protected FilterSpans.AcceptStatus accept(BLSpans candidate) throws IOException {
-        if (mode == MatchInfo.SpanMode.SOURCE & in.getRelationInfo().isRoot()) {
+        if (mode == RelationInfo.SpanMode.SOURCE & in.getRelationInfo().isRoot()) {
             // Need source, but this has no source
             return FilterSpans.AcceptStatus.NO;
         }
@@ -90,7 +90,7 @@ class SpansRelationSpanAdjust extends BLFilterSpans<BLSpans> {
             startAdjusted = endAdjusted = NO_MORE_POSITIONS;
         } else if (atFirstInCurrentDoc || startPos < 0) {
             startAdjusted = endAdjusted = -1;
-        } else if (mode == MatchInfo.SpanMode.ALL_SPANS) {
+        } else if (mode == RelationInfo.SpanMode.ALL_SPANS) {
             // We need all match info because we want the full span including all matched relations
             if (matchInfo == null)
                 matchInfo = new MatchInfo[context.numberOfMatchInfos()];
@@ -111,8 +111,8 @@ class SpansRelationSpanAdjust extends BLFilterSpans<BLSpans> {
             }
             if (startAdjusted == Integer.MAX_VALUE) {
                 // Weird, no relations matched; use the original span (this should never happen though)
-                startAdjusted = in.getRelationInfo().spanStart(MatchInfo.SpanMode.FULL_SPAN);
-                endAdjusted = in.getRelationInfo().spanEnd(MatchInfo.SpanMode.FULL_SPAN);
+                startAdjusted = in.getRelationInfo().spanStart(RelationInfo.SpanMode.FULL_SPAN);
+                endAdjusted = in.getRelationInfo().spanEnd(RelationInfo.SpanMode.FULL_SPAN);
             }
         } else {
             startAdjusted = in.getRelationInfo().spanStart(mode);
@@ -139,7 +139,7 @@ class SpansRelationSpanAdjust extends BLFilterSpans<BLSpans> {
             // Our cached hit is the one we want.
             return nextStartPosition();
         }
-        if (mode != MatchInfo.SpanMode.FULL_SPAN) {
+        if (mode != RelationInfo.SpanMode.FULL_SPAN) {
             // We can't skip because the spans we produce are not guaranteed to be sorted by start position.
             // Call the naive implementation.
             if (BLSpans.naiveAdvanceStartPosition(this, target) == NO_MORE_POSITIONS) {
