+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansRelations.java
@@ -27,7 +27,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
     private int endPos = NOT_YET_NEXTED;
 
     /** Source and target for this relation */
-    private final MatchInfo relationInfo = new MatchInfo();
+    private final RelationInfo relationInfo = new RelationInfo();
 
     /** If true, we have to skip the primary value indicator in the payload (see PayloadUtils) */
     private final boolean payloadIndicatesPrimaryValues;
@@ -36,7 +36,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
     private final Direction direction;
 
     /** What span to return for the relations found */
-    private final MatchInfo.SpanMode spanMode;
+    private final RelationInfo.SpanMode spanMode;
 
     /** Group number where we'll capture our relation info */
     private int groupIndex;
@@ -56,7 +56,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
      * @param payloadIndicatesPrimaryValues whether or not there's "is primary value" indicators in the payloads
      */
     public SpansRelations(String relationType, BLSpans relationsMatches,
-            boolean payloadIndicatesPrimaryValues, Direction direction, MatchInfo.SpanMode spanMode) {
+            boolean payloadIndicatesPrimaryValues, Direction direction, RelationInfo.SpanMode spanMode) {
         super(relationsMatches, SpanQueryRelations.createGuarantees(relationsMatches.guarantees(), direction, spanMode));
         this.relationType = relationType;
         this.payloadIndicatesPrimaryValues = payloadIndicatesPrimaryValues;
@@ -83,7 +83,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
         return true;
     }
 
-    public MatchInfo getRelationInfo() {
+    public RelationInfo getRelationInfo() {
         return relationInfo;
     }
 
@@ -130,7 +130,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
             return nextStartPosition();
         }
         if (direction == Direction.FORWARD &&
-                (spanMode == MatchInfo.SpanMode.FULL_SPAN || spanMode == MatchInfo.SpanMode.SOURCE)) {
+                (spanMode == RelationInfo.SpanMode.FULL_SPAN || spanMode == RelationInfo.SpanMode.SOURCE)) {
             // We know our spans will be in order, so we can use the more efficient advanceStartPosition()
             super.advanceStartPosition(target);
             if (startPos == NO_MORE_POSITIONS) {
@@ -162,7 +162,7 @@ class SpansRelations extends BLFilterSpans<BLSpans> {
     @Override
     protected FilterSpans.AcceptStatus accept(BLSpans candidate) throws IOException {
         fetchRelationInfo(); // decode the payload
-        if (relationInfo.isRoot() && spanMode == MatchInfo.SpanMode.SOURCE) {
+        if (relationInfo.isRoot() && spanMode == RelationInfo.SpanMode.SOURCE) {
             // Root relations have no source
             return FilterSpans.AcceptStatus.NO;
         }
