+++ engine/src/main/java/nl/inl/blacklab/search/lucene/SpansRepetition.java
@@ -235,7 +235,7 @@ class SpansRepetition extends BLFilterDocsSpans<SpansInBucketsPerDocumentWithSta
     }
 
     @Override
-    public MatchInfo getRelationInfo() {
+    public RelationInfo getRelationInfo() {
         return in.getRelationInfo(matchStartIndex);
     }
 }
