+++ engine/src/main/java/nl/inl/blacklab/search/results/Hits.java
@@ -15,6 +15,7 @@ import nl.inl.blacklab.search.indexmetadata.Annotation;
 import nl.inl.blacklab.search.indexmetadata.MatchSensitivity;
 import nl.inl.blacklab.search.lucene.BLSpanQuery;
 import nl.inl.blacklab.search.lucene.MatchInfo;
+import nl.inl.blacklab.search.lucene.RelationInfo;
 
 /**
  * A list of hits, optionally with captured groups.
