+++ text-pattern/src/main/java/nl/inl/blacklab/search/extensions/XFRelations.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import nl.inl.blacklab.search.QueryExecutionContext;
 import nl.inl.blacklab.search.lucene.BLSpanQuery;
-import nl.inl.blacklab.search.lucene.MatchInfo;
+import nl.inl.blacklab.search.lucene.RelationInfo;
 import nl.inl.blacklab.search.lucene.SpanQueryRelationSpanAdjust;
 import nl.inl.blacklab.search.lucene.SpanQueryRelations;
 import nl.inl.blacklab.search.results.QueryInfo;
@@ -18,7 +18,7 @@ public class XFRelations implements ExtensionFunctionClass {
 
     private static BLSpanQuery rel(QueryInfo queryInfo, QueryExecutionContext context, List<Object> args) {
         String relationType = (String) args.get(0);
-        MatchInfo.SpanMode spanMode = MatchInfo.SpanMode.fromCode((String)args.get(1));
+        RelationInfo.SpanMode spanMode = RelationInfo.SpanMode.fromCode((String)args.get(1));
         SpanQueryRelations.Direction direction = SpanQueryRelations.Direction.fromCode((String) args.get(2));
         String field = context.withRelationAnnotation().luceneField();
         return new SpanQueryRelations(queryInfo, field, relationType, (Map<String, String> )null, direction,
@@ -27,7 +27,7 @@ public class XFRelations implements ExtensionFunctionClass {
 
     private static BLSpanQuery rspan(QueryInfo queryInfo, QueryExecutionContext context, List<Object> args) {
         BLSpanQuery relations = (BLSpanQuery) args.get(0);
-        MatchInfo.SpanMode mode = MatchInfo.SpanMode.fromCode((String)args.get(1));
+        RelationInfo.SpanMode mode = RelationInfo.SpanMode.fromCode((String)args.get(1));
         return new SpanQueryRelationSpanAdjust(relations, mode);
     }
